♻️ refactor(server): 리소스 핸들러 파일 이름 변경
- 파일 이름을 resources.py에서 handlers.py로 변경
- 모듈 이름을 더 명확하게 반영
